Updated to allow multi-target sync
rscs2 now allows for multiple targets to be synced in one instance of
the script. For example, with directory options "school" and "work" and
"dotfiles", you could simply run `rscs school work` to sync both school
and work but not dotfiles. Before this commit, this would have needed to
be `rscs school` and `rscs work` separately.